AI/Image_Detection/requirements.txt
tensorflow==2.14.0
numpy<2.0.0
Pillow>=10.0.0